divide task and statistical metrics
