+++ tests/unit/test_task_metrics.py
@@ -7,9 +7,8 @@ import math
 import pytest
 import numpy as np
 
-from armory.utils import metrics
+from armory.metrics import task
 
-# Mark all tests in this file as `unit`
 pytestmark = pytest.mark.unit
 
 
@@ -20,8 +19,8 @@ def test_entailment():
     """
     Slow due to 1 GB file download and multiple model predictions
     """
-    metric = metrics.Entailment()
-    metric_repeat = metrics.Entailment()
+    metric = task.Entailment()
+    metric_repeat = task.Entailment()
     assert metric.model is metric_repeat.model
 
     from armory.attacks.librispeech_target_labels import (
@@ -62,16 +61,16 @@ def test_entailment():
 def test_total_entailment(caplog):
     for invalid_results in (["invalid name"], [-1], [3], [None]):
         with pytest.raises(ValueError):
-            metrics.total_entailment(invalid_results)
+            task.total_entailment(invalid_results)
 
-    metrics.total_entailment([0, 1, 2])
+    task.total_entailment([0, 1, 2])
     assert "Entailment outputs are (0, 1, 2) ints, not strings, mapping" in caplog.text
 
-    results = metrics.total_entailment(["contradiction"])
+    results = task.total_entailment(["contradiction"])
     assert results == dict(contradiction=1, neutral=0, entailment=0)
     assert isinstance(results, dict)
 
-    results = metrics.total_entailment([0, 1, 1, "neutral", "entailment", "entailment"])
+    results = task.total_entailment([0, 1, 1, "neutral", "entailment", "entailment"])
     assert results == dict(contradiction=1, neutral=3, entailment=2)
     assert sum(results.values()) == 6
 
@@ -79,7 +78,7 @@ def test_total_entailment(caplog):
 def test_tpr_fpr():
     actual_conditions = [0] * 10 + [1] * 10
     predicted_conditions = [0] * 4 + [1] * 6 + [0] * 3 + [1] * 7
-    results = metrics.tpr_fpr(actual_conditions, predicted_conditions)
+    results = task.tpr_fpr(actual_conditions, predicted_conditions)
     for k, v in [
         ("true_positives", 7),
         ("true_negatives", 4),
@@ -93,7 +92,7 @@ def test_tpr_fpr():
     ]:
         assert results[k] == v, "k"
 
-    results = metrics.tpr_fpr([], [])
+    results = task.tpr_fpr([], [])
     for k, v in [
         ("true_positives", 0),
         ("true_negatives", 0),
@@ -111,40 +110,40 @@ def test_tpr_fpr():
             assert results[k] == v, f"{k}"
 
     with pytest.raises(ValueError):
-        metrics.tpr_fpr(actual_conditions, [predicted_conditions])
+        task.tpr_fpr(actual_conditions, [predicted_conditions])
     for array in 0, [[0, 1], [1, 0]]:
         with pytest.raises(ValueError):
-            metrics.tpr_fpr(array, array)
+            task.tpr_fpr(array, array)
 
 
 def test_abstains():
     y = [0, 1, 2, 3, 4]
     y_pred = np.zeros((5, 10))
     y_pred[1, 1] = 1.0
-    assert metrics.abstains(y, y_pred) == [1, 0, 1, 1, 1]
+    assert task.abstains(y, y_pred) == [1, 0, 1, 1, 1]
     for wrong_dim in np.zeros(5), np.zeros((5, 10, 10)):
         with pytest.raises(ValueError):
-            metrics.abstains(y, np.zeros(5))
+            task.abstains(y, np.zeros(5))
 
 
 def test_categorical_accuracy():
     y = [0, 1, 2, 3, 4]
     y_pred = [0, 1, 2, 2, 3]
-    assert metrics.categorical_accuracy(y, y_pred) == [1, 1, 1, 0, 0]
-    assert metrics.categorical_accuracy(y, np.eye(5)) == [1] * 5
+    assert task.categorical_accuracy(y, y_pred) == [1, 1, 1, 0, 0]
+    assert task.categorical_accuracy(y, np.eye(5)) == [1] * 5
     with pytest.raises(ValueError):
-        metrics.categorical_accuracy(y, [[y_pred]])
+        task.categorical_accuracy(y, [[y_pred]])
 
 
 def test_top_n_categorical_accuracy():
     y = [0, 1, 2, 3, 4]
     y_pred = [0, 1, 2, 2, 3]
-    assert metrics.categorical_accuracy(
-        y, y_pred
-    ) == metrics.top_n_categorical_accuracy(y, y_pred, 1)
+    assert task.categorical_accuracy(y, y_pred) == task.top_n_categorical_accuracy(
+        y, y_pred, 1
+    )
     y = [2, 0]
     y_pred = [[0.1, 0.4, 0.2, 0.2, 0.1, 0.1], [0.1, 0.4, 0.2, 0.2, 0.1, 0.1]]
-    assert metrics.top_n_categorical_accuracy(y, y_pred, 3) == [1, 0]
+    assert task.top_n_categorical_accuracy(y, y_pred, 3) == [1, 0]
 
 
 def test_mAP():
@@ -156,7 +155,7 @@ def test_mAP():
         "scores": np.array([0.8, 0.8]),
     }
 
-    ap_per_class = metrics.object_detection_AP_per_class([labels], [preds])
+    ap_per_class = task.object_detection_AP_per_class([labels], [preds])
     assert ap_per_class[9] == 0
     assert ap_per_class[2] >= 0.99
 
@@ -191,7 +190,7 @@ def test_object_detection_metrics():
         misclassification_rate_per_img,
         disappearance_rate_per_img,
         hallucinations_per_img,
-    ) = metrics._object_detection_get_tpr_mr_dr_hr(
+    ) = task._object_detection_get_tpr_mr_dr_hr(
         y, y_pred, score_threshold=score_threshold, iou_threshold=iou_threshold
     )
     for rate_per_img in [
@@ -208,7 +207,7 @@ def test_object_detection_metrics():
 
 
 def test_video_tracking_metrics():
-    # Recall that first box isn't counted towards metrics. Second box has IoU of 1.0,
+    # Recall that first box isn't counted towards task. Second box has IoU of 1.0,
     # third box has IoU of 0
     y = [
         {
@@ -225,8 +224,8 @@ def test_video_tracking_metrics():
         }
     ]
 
-    mean_iou = metrics.video_tracking_mean_iou(y, y_pred)
-    mean_success = metrics.video_tracking_mean_success_rate(y, y_pred)
+    mean_iou = task.video_tracking_mean_iou(y, y_pred)
+    mean_success = task.video_tracking_mean_success_rate(y, y_pred)
 
     for result in [mean_success, mean_iou]:
         assert isinstance(result, list)
